requirements.txt
Django
Pillow<7.0.0,>=4.0.0
Willow
Wand
Unidecode
asgiref
beautifulsoup4<4.6.1,>=4.5.1
certifi
chardet
django-cogwheels
django-colorfield
django-debug-toolbar
django-modelcluster
django-pipeline
django-taggit
django-treebeard
djangorestframework
draftjs-exporter<3.0,>=2.1.5
idna
l18n
pip
psycopg2-binary
pytz
requests
setuptools
six
sqlparse
urllib3
wagtail==2.9
wagtailmenus